requirements.txt
# Dependencias principales para RAG en LangChain (Python)
langchain>=0.2,<1.0
langchain-community>=0.2,<1.0
langchain-openai>=0.1.0
langchain-text-splitters>=0.3
faiss-cpu>=1.7.4
pypdf>=4.0.0
python-dotenv>=1.0.0
fastapi>=0.110
uvicorn[standard]>=0.27
pytest>=8.0.0
